+++ activerecord/lib/active_record/associations/join_dependency.rb
@@ -28,7 +28,7 @@ module ActiveRecord
       #    associations #=> [:appointments]
       #    joins #=>  []
       #
-      def initialize(base, associations, joins)
+      def initialize(base, associations, joins, join_type=nil)
         @base_klass    = base
         @table_joins   = joins
         @join_parts    = [JoinBase.new(base)]
@@ -36,7 +36,7 @@ module ActiveRecord
         @reflections   = []
         @alias_tracker = AliasTracker.new(base.connection, joins)
         @alias_tracker.aliased_name_for(base.table_name) # Updates the count for base.table_name to 1
-        build(associations)
+        build(associations, @join_parts.last, join_type)
       end
 
       def graft(*associations)