+++ activerecord/lib/active_record/relation.rb
@@ -6,8 +6,8 @@ module ActiveRecord
     JoinOperation = Struct.new(:relation, :join_class, :on)
 
     MULTI_VALUE_METHODS  = [:includes, :eager_load, :preload, :select, :group,
-                            :order, :joins, :where, :having, :bind, :references,
-                            :extending]
+                            :order, :joins, :outer_joins, :where, :having,
+                            :bind, :references, :extending]
 
     SINGLE_VALUE_METHODS = [:limit, :offset, :lock, :readonly, :from, :reordering,
                             :reverse_order, :distinct, :create_with, :uniq]